moves simulatedPlayer to debug module
